WIP: ENH: ViewImage available as non-static class
`ViewImage` class can be called using its static function `View` if
the user does not need to keep track of the `ViewImage` object.
However, in some case, the user needs to keep track of the object
and can now use the class through the typical ITK pointer interface.
Using this interface, the user has the extra option to set
the render window interactor.